// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/permissionsResponse.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * The version of the OpenAPI document: 3.14.0-SNAPSHOT
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Map of Permissions.
 */
export interface PermissionsResponse { 
    METADATA?: Array<string>;
    DOCUMENTATION?: Array<string>;
    DEFINITION?: Array<string>;
    PLAN?: Array<string>;
    SUBSCRIPTION?: Array<string>;
    MEMBER?: Array<string>;
    ANALYTICS?: Array<string>;
    EVENT?: Array<string>;
    HEALTH?: Array<string>;
    LOG?: Array<string>;
    GATEWAY_DEFINITION?: Array<string>;
    AUDIT?: Array<string>;
    RATING?: Array<string>;
    RATING_ANSWER?: Array<string>;
    NOTIFICATION?: Array<string>;
    MESSAGE?: Array<string>;
    ALERT?: Array<string>;
    RESPONSE_TEMPLATES?: Array<string>;
    REVIEWS?: Array<string>;
    QUALITY_RULE?: Array<string>;
}